requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
diff-match-patch==20241021
Django==5.1.2
django-ckeditor==6.7.2
django-import-export==4.3.3
django-ipware==7.0.1
django-js-asset==3.0.1
django-widget-tweaks==1.5.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.0.0
jellyfish==1.1.3
Levenshtein==0.26.1
packaging==24.2
pdfkit==1.0.0
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyotp==2.9.0
pytest==8.3.4
python-dotenv==1.0.1
python-ipware==3.0.0
python-Levenshtein==0.26.1
RapidFuzz==3.12.1
razorpay==1.4.2
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.1
tablib==3.7.0
urllib3==2.2.3
